+++ src/standalone/Snow/boundary_fluxes.jl
@@ -127,17 +127,10 @@ function snow_boundary_fluxes!(
         P_snow +
         (P_liq + p.snow.turbulent_fluxes.vapor_flux - p.snow.water_runoff) *
         p.snow.snow_cover_fraction
-    residual_melt_flux = get_residual_melt_flux(
-        model.parameters.surf_temp,
-        Y,
-        p,
-        model.parameters.earth_param_set,
-    )
-
     @. p.snow.liquid_water_flux =
         (
             P_liq + p.snow.turbulent_fluxes.vapor_flux * p.snow.q_l -
-            p.snow.water_runoff + residual_melt_flux
+            p.snow.water_runoff
         ) * p.snow.snow_cover_fraction
 
     e_flux_falling_snow =